backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0
pymongo==4.6.0
bcrypt==4.1.2
python-jose[cryptography]==3.3.0
passlib==1.7.4
pydantic==2.5.0
email-validator==2.1.0
aiosmtplib==2.0.2
aioimaplib==1.0.1
python-email-utils==0.1.2
httpx==0.25.2
Pillow==10.1.0
python-magic==0.4.27